Add tests with local declarations
