+++ analyzers/tests/SonarAnalyzer.UnitTest/Rules/Utilities/TokenTypeAnalyzerTest.Classifier.MS.cs
@@ -2,22 +2,5 @@
 
 public partial class TokenTypeAnalyzerTest
 {
-    [DataTestMethod]
-    [DataRow("[t:Exception] ex;")]        // -model
-    [DataRow("[u:System].Exception ex;")] // +model
-    [DataRow("[t:List]<[t:Exception]> ex;")] // -model
-    [DataRow("List<[u:System].Exception> ex;")] // +model
-    public void IdentifierToken_LocalDeclaration(string declaration, bool allowSemanticModel = true) =>
-        ClassifierTestHarness.AssertTokenTypes(
-        $$"""
-        using System;
-        using System.Collections.Generic;
-        public class Test
-        {
-            public void M()
-            {
-                {{declaration}}
-            }
-        }
-        """, allowSemanticModel);
+
 }
